# docs/openapi/panel-api.v1.yaml
openapi: 3.0.3
info:
  title: Lenker Panel API
  version: 0.1.0-draft
  description: Draft OpenAPI contract for the currently implemented Lenker panel-api MVP v0.1 backend surface.
servers:
  - url: http://localhost:8080
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/subscription-templates:
    get:
      tags: [SubscriptionTemplates]
      summary: List subscription templates
      responses:
        "200":
          description: Subscription templates list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionTemplateListEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [SubscriptionTemplates]
      summary: Create a subscription template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionTemplateRequest"
      responses:
        "201":
          description: Subscription template created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionTemplateEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/subscription-templates/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags: [SubscriptionTemplates]
      summary: Get a subscription template
      responses:
        "200":
          description: Subscription template details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionTemplateEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      tags: [SubscriptionTemplates]
      summary: Update a subscription template
      description: Only user-created templates can be updated. System templates return 400.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubscriptionTemplateRequest"
      responses:
        "200":
          description: Subscription template updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionTemplateEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [SubscriptionTemplates]
      summary: Delete a subscription template
      description: Only user-created templates can be deleted. System templates return 400.
      responses:
        "200":
          description: Subscription template deleted.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/subscription-templates/{id}/create-subscription:
    parameters:
      - $ref: "#/components/parameters/ID"
    post:
      tags: [SubscriptionTemplates]
      summary: Create a subscription from a template
      description: Creates a new subscription using the template's plan_id.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionFromTemplateRequest"
      responses:
        "201":
          description: Subscription created from template.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: session-token
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource identifier.
  responses:
    BadRequest:
      description: Invalid request body or validation failure.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Missing or invalid admin session token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: Resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    InternalError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [data]
      properties:
        data:
          description: Successful response payload.
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/APIError"
    APIError:
      type: object
      required: [code, message]
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
    ErrorCode:
      type: string
      enum:
        - bad_request
        - unauthorized
        - not_found
        - validation_error
        - storage_error
        - internal_error
        - invalid_credentials
        - inactive_admin
        - invalid_bootstrap_token
        - expired_bootstrap_token
        - bootstrap_token_used
...
    Subscription:
      type: object
      required:
        - id
        - user_id
        - plan_id
        - status
        - starts_at
        - expires_at
        - traffic_used_bytes
        - device_limit
      properties:
        id:
          type: string
        user_id:
          type: string
        plan_id:
          type: string
        status:
          type: string
          enum: [active, expired, suspended]
        starts_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
        traffic_limit_bytes:
          type: integer
          format: int64
          minimum: 1
          nullable: true
        traffic_used_bytes:
          type: integer
          format: int64
          minimum: 0
        device_limit:
          type: integer
          minimum: 1
        preferred_region:
          type: string
          nullable: true
    SubscriptionEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/Subscription"
...
    SubscriptionTemplate:
      type: object
      required: [id, name, config, is_system, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        plan_id:
          type: string
          format: uuid
          nullable: true
        config:
          type: object
          properties:
            duration_days:
              type: integer
            traffic_limit_bytes:
              type: integer
              format: int64
              nullable: true
            device_limit:
              type: integer
        is_system:
          type: boolean
        created_at:
          type: string
          format: date-time
    SubscriptionTemplateEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/SubscriptionTemplate"
    SubscriptionTemplateListEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/SubscriptionTemplate"
    CreateSubscriptionTemplateRequest:
      type: object
      required: [name, config]
      properties:
        name:
          type: string
        description:
          type: string
        plan_id:
          type: string
          format: uuid
        config:
          type: object
          required: [duration_days, device_limit]
          properties:
            duration_days:
              type: integer
              minimum: 1
            traffic_limit_bytes:
              type: integer
              format: int64
              nullable: true
            device_limit:
              type: integer
              minimum: 1
    UpdateSubscriptionTemplateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        plan_id:
          type: string
          format: uuid
        config:
          type: object
          properties:
            duration_days:
              type: integer
              minimum: 1
            traffic_limit_bytes:
              type: integer
              format: int64
              nullable: true
            device_limit:
              type: integer
              minimum: 1
    CreateSubscriptionFromTemplateRequest:
      type: object
      required: [user_id]
      properties:
        user_id:
          type: string
        preferred_region:
          type: string
          nullable: true